Cleaned up ResultSection component, updated VolcanoPlot data fetch
